mise-asi/requirements.txt
uagents
flask
flask-cors
supabase
python-dotenv
requests
openai
